refactor: rename tokenAuthMiddleware to authMiddleware, gate app-wide, use errorHandler
