+++ backend/routes/video/video.ts
@@ -10,7 +10,6 @@ import ArchiveHookPayloadSchema from './schemas/ArchiveHookPayload.schema';
 import { VideoSessionDetails } from '@common/types';
 import { assertResult } from '@api-lib/executions';
 import getSessionStorageService from '../../sessionStorageService';
-import { tokenAuthMiddleware } from '../../middleware';
 import { CaptionsStatus } from './types';
 
 const videoRouter = Router();
@@ -177,13 +176,6 @@ videoRouter.post(
 );
 
 // #endregion ------------------------------------------------------------------------
-
-/**
- * OIDC token validation on the protected session/token endpoints.
- * A no-op unless AUTH_ENABLED=true (see backend/middleware/tokenAuthMiddleware).
- */
-videoRouter.post('/createSession', tokenAuthMiddleware);
-videoRouter.post('/joinSession', tokenAuthMiddleware);
 
 videoRouter.use(videoHandler);
