+++ backend/tests/authMiddleware.test.ts
@@ -10,7 +10,7 @@ process.env.OIDC_CLIENT_ID = 'test-client-id';
 process.env.OIDC_ISSUER_URL = 'https://example.com';
 const startServer = (await import('../server')).default;
 
-describe('tokenAuthMiddleware, mounted on the real app', () => {
+describe('authMiddleware, mounted on the real app', () => {
   let server: Server;
 
   beforeAll(async () => {
@@ -25,13 +25,13 @@ describe('tokenAuthMiddleware, mounted on the real app', () => {
     const res = await request(server).post('/v2/createSession').send({});
 
     expect(res.statusCode).toEqual(401);
-    expect(res.body).toHaveProperty('error');
+    expect(res.body).toHaveProperty('message');
   });
 
   it('rejects an unauthenticated POST /v2/joinSession with a real 401 response', async () => {
     const res = await request(server).post('/v2/joinSession').send({});
 
     expect(res.statusCode).toEqual(401);
-    expect(res.body).toHaveProperty('error');
+    expect(res.body).toHaveProperty('message');
   });
 });
